Rev 148

+++ acuity-api/src/main/java/com/acuity/rs/api/RSBoundingBox.java
@@ -5,5 +5,5 @@ package com.acuity.rs.api;
 
 public interface RSBoundingBox extends RSNode {
 
-	void invokeDraw(byte var0);
+	void invokeDraw(int var0);
 }
